pull out TokenRangeFetcher
