+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/cassandra/backup/CqlCluster.java
@@ -17,39 +17,18 @@
 package com.palantir.atlasdb.cassandra.backup;
 
 import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.ConsistencyLevel;
-import com.datastax.driver.core.KeyspaceMetadata;
-import com.datastax.driver.core.Statement;
-import com.datastax.driver.core.TableMetadata;
-import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.RangeSet;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
-import com.palantir.atlasdb.cassandra.CassandraServersConfigs;
-import com.palantir.atlasdb.cassandra.CassandraServersConfigs.CqlCapableConfig;
 import com.palantir.atlasdb.cassandra.backup.transaction.TransactionsTableInteraction;
-import com.palantir.atlasdb.keyvalue.cassandra.CassandraConstants;
 import com.palantir.atlasdb.keyvalue.cassandra.LightweightOppToken;
 import com.palantir.atlasdb.keyvalue.cassandra.async.client.creation.ClusterFactory;
-import com.palantir.logsafe.SafeArg;
-import com.palantir.logsafe.exceptions.SafeIllegalStateException;
-import com.palantir.logsafe.logger.SafeLogger;
-import com.palantir.logsafe.logger.SafeLoggerFactory;
 import java.io.Closeable;
 import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
 
 public final class CqlCluster implements Closeable {
-    private static final SafeLogger log = SafeLoggerFactory.get(CqlCluster.class);
-
-    private static final int LONG_READ_TIMEOUT_MS = (int) TimeUnit.MINUTES.toMillis(2);
-    // reduce this from default because we run RepairTableTask across N keyspaces at the same time
-    private static final int SELECT_FETCH_SIZE = 1_000;
-
     private final Cluster cluster;
     private final CassandraKeyValueServiceConfig config;
 
@@ -69,36 +48,9 @@ public final class CqlCluster implements Closeable {
         cluster.close();
     }
 
-    private static Set<InetSocketAddress> getHosts(CassandraKeyValueServiceConfig config) {
-        return CassandraServersConfigs.getCqlCapableConfigIfValid(config)
-                .map(CqlCapableConfig::cqlHosts)
-                .orElseThrow(() -> new SafeIllegalStateException("Attempting to get token ranges with thrift config!"));
-    }
-
     public Map<InetSocketAddress, RangeSet<LightweightOppToken>> getTokenRanges(String tableName) {
         try (CqlSession session = new CqlSession(cluster.connect())) {
-            String keyspaceName = config.getKeyspaceOrThrow();
-            CqlMetadata metadata = session.getMetadata();
-            KeyspaceMetadata keyspace = metadata.getKeyspaceMetadata(keyspaceName);
-            TableMetadata tableMetadata = keyspace.getTable(tableName);
-            Set<LightweightOppToken> partitionTokens = getPartitionTokens(session, tableMetadata);
-            Map<InetSocketAddress, RangeSet<LightweightOppToken>> tokenRangesByNode =
-                    ClusterMetadataUtils.getTokenMapping(getHosts(config), metadata, keyspaceName, partitionTokens);
-
-            if (!partitionTokens.isEmpty() && log.isDebugEnabled()) {
-                int numTokenRanges = tokenRangesByNode.values().stream()
-                        .mapToInt(ranges -> ranges.asRanges().size())
-                        .sum();
-
-                log.debug(
-                        "Identified token ranges requiring repair",
-                        SafeArg.of("keyspace", keyspace),
-                        SafeArg.of("table", tableName),
-                        SafeArg.of("numPartitionKeys", partitionTokens.size()),
-                        SafeArg.of("numTokenRanges", numTokenRanges));
-            }
-
-            return tokenRangesByNode;
+            return new TokenRangeFetcher(session, config).getTokenRange(tableName);
         }
     }
 
@@ -108,23 +60,5 @@ public final class CqlCluster implements Closeable {
             return new RepairRangeFetcher(session, config)
                     .getTransactionTableRangesForRepair(transactionsTableInteractions);
         }
-    }
-
-    private static Set<LightweightOppToken> getPartitionTokens(CqlSession session, TableMetadata tableMetadata) {
-        return session.retrieveRowKeysAtConsistencyAll(createSelectStatements(tableMetadata));
-    }
-
-    private static List<Statement> createSelectStatements(TableMetadata table) {
-        return ImmutableList.of(createSelectStatement(table));
-    }
-
-    private static Statement createSelectStatement(TableMetadata table) {
-        return QueryBuilder.select(CassandraConstants.ROW)
-                // only returns the column that we need instead of all of them, otherwise we get a timeout
-                .distinct()
-                .from(table)
-                .setConsistencyLevel(ConsistencyLevel.ALL)
-                .setFetchSize(SELECT_FETCH_SIZE)
-                .setReadTimeoutMillis(LONG_READ_TIMEOUT_MS);
     }
 }
